+++ ClickerHeroesTrackerWebsite/ClickerHeroesTrackerWebsite/ViewModels/Dashboard/ProgressData.cs
@@ -28,18 +28,18 @@ namespace ClickerHeroesTrackerWebsite.Models.Dashboard
         {
             this.UserSettings = userSettings;
 
-            this.OptimalLevelData = new SortedDictionary<DateTime, long>();
-            this.SoulsPerHourData = new SortedDictionary<DateTime, long>();
-            this.TitanDamageData = new SortedDictionary<DateTime, long>();
-            this.SoulsSpentData = new SortedDictionary<DateTime, long>();
+            this.OptimalLevelData = new SortedDictionary<DateTime, double>();
+            this.SoulsPerHourData = new SortedDictionary<DateTime, double>();
+            this.TitanDamageData = new SortedDictionary<DateTime, double>();
+            this.SoulsSpentData = new SortedDictionary<DateTime, double>();
 
             while (reader.Read())
             {
                 var uploadTime = Convert.ToDateTime(reader["UploadTime"]);
                 this.OptimalLevelData.AddOrUpdate(uploadTime, Convert.ToInt64(reader["OptimalLevel"]));
                 this.SoulsPerHourData.AddOrUpdate(uploadTime, Convert.ToInt64(reader["SoulsPerHour"]));
-                this.TitanDamageData.AddOrUpdate(uploadTime, Convert.ToInt64(reader["TitanDamage"]));
-                this.SoulsSpentData.AddOrUpdate(uploadTime, Convert.ToInt64(reader["SoulsSpent"]));
+                this.TitanDamageData.AddOrUpdate(uploadTime, Convert.ToDouble(reader["TitanDamage"]));
+                this.SoulsSpentData.AddOrUpdate(uploadTime, Convert.ToDouble(reader["SoulsSpent"]));
             }
 
             if (!reader.NextResult())
@@ -47,13 +47,13 @@ namespace ClickerHeroesTrackerWebsite.Models.Dashboard
                 return;
             }
 
-            this.AncientLevelData = new SortedDictionary<Ancient, IDictionary<DateTime, long>>(AncientComparer.Instance);
+            this.AncientLevelData = new SortedDictionary<Ancient, IDictionary<DateTime, double>>(AncientComparer.Instance);
             while (reader.Read())
             {
                 // UploadTime, AncientId, Level
                 var uploadTime = Convert.ToDateTime(reader["UploadTime"]);
                 var ancientId = Convert.ToInt32(reader["AncientId"]);
-                var level = Convert.ToInt64(reader["Level"]);
+                var level = Convert.ToDouble(reader["Level"]);
 
                 Ancient ancient;
                 if (!gameData.Ancients.TryGetValue(ancientId, out ancient))
@@ -68,10 +68,10 @@ namespace ClickerHeroesTrackerWebsite.Models.Dashboard
                     continue;
                 }
 
-                IDictionary<DateTime, long> levelData;
+                IDictionary<DateTime, double> levelData;
                 if (!this.AncientLevelData.TryGetValue(ancient, out levelData))
                 {
-                    levelData = new SortedDictionary<DateTime, long>();
+                    levelData = new SortedDictionary<DateTime, double>();
                     this.AncientLevelData.Add(ancient, levelData);
                 }
 
@@ -94,27 +94,27 @@ namespace ClickerHeroesTrackerWebsite.Models.Dashboard
         /// <summary>
         /// Gets the optimal level data, keyed on upload time.
         /// </summary>
-        public IDictionary<DateTime, long> OptimalLevelData { get; }
+        public IDictionary<DateTime, double> OptimalLevelData { get; }
 
         /// <summary>
         /// Gets the souls per hour data, keyed on upload time.
         /// </summary>
-        public IDictionary<DateTime, long> SoulsPerHourData { get; }
+        public IDictionary<DateTime, double> SoulsPerHourData { get; }
 
         /// <summary>
         /// Gets the titan damage data, keyed on upload time.
         /// </summary>
-        public IDictionary<DateTime, long> TitanDamageData { get; }
+        public IDictionary<DateTime, double> TitanDamageData { get; }
 
         /// <summary>
         /// Gets the souls spent data, keyed on upload time.
         /// </summary>
-        public IDictionary<DateTime, long> SoulsSpentData { get; }
+        public IDictionary<DateTime, double> SoulsSpentData { get; }
 
         /// <summary>
         /// Gets the ancient level data, keyed on ancient. The inner dictionaries are keyed on upload time.
         /// </summary>
-        public IDictionary<Ancient, IDictionary<DateTime, long>> AncientLevelData { get; }
+        public IDictionary<Ancient, IDictionary<DateTime, double>> AncientLevelData { get; }
 
         private class AncientComparer : IComparer<Ancient>
         {
